requirements.txt
a2wsgi==1.10.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
attrs==25.3.0
bentoml==1.4.15
cattrs==23.1.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-option-group==0.5.7
cloudpickle==3.1.1
cryptography==45.0.3
Deprecated==1.2.18
ecdsa==0.19.1
fastapi==0.115.12
frozenlist==1.6.0
fs==2.4.16
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.7.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.0
kantoku==0.18.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
numpy==2.2.6
nvidia-ml-py==12.575.51
opentelemetry-api==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-aiohttp-client==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pip-requirements-parser==32.0.1
pluggy==1.6.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
questionary==2.1.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
simple-di==0.1.5
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
tomli_w==1.2.0
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uv==0.7.6
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.22.0
